\name{Passport}
\alias{passport}
\alias{chrom}
\alias{Fst}
\title{
Passport data
}
\description{
Germplasm collection passport information
}
\usage{
data(pass)
}
\details{
Passport information include quantitative measures, source, domestication, and a cluster classification based on Nei's standard genetic distance and Ward's D algorithm of hierarchical clustering. The dataset also includes Fst values for each SNP, and a vector informing the number of markers in each chromosome. Passport information was provided by Randall Nelson.
}
\examples{
\dontrun{
data(pass)
}
}
