\name{SGC-package}
\alias{SGC-package}
\docType{package}
\title{
Soybean Germplasm Collection dataset
}
\description{
Data used by Xavier et al. (2018) <doi.org/10.1017/S1479262118000102>. USDA soybean germplasm collection dataset, including genotypic information, passport data, clusters and FST values.
}
\details{
\tabular{ll}{
Package: \tab SGC\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2019-03-12\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, Rima Thapa, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}